Drop shaow DOM from nested components
