+++ src/api-viewer-demo-events.ts
@@ -2,12 +2,10 @@ import {
   LitElement,
   html,
   customElement,
-  css,
   property,
   TemplateResult
 } from 'lit-element';
 import { cache } from 'lit-html/directives/cache.js';
-import buttonStyle from './lib/button-style.js';
 
 interface EventDetail {
   value: string | number | boolean | null | undefined;
@@ -25,7 +23,7 @@ const renderEvents = (log: CustomEvent[]): TemplateResult => {
   return html`
     ${log.map(e => {
       return html`
-        <p>
+        <p class="event">
           event: "${e.type}". detail: ${renderDetail(e.detail)}
         </p>
       `;
@@ -38,35 +36,8 @@ export class ApiViewerDemoEvents extends LitElement {
   @property({ attribute: false, hasChanged: () => true })
   log: CustomEvent[] = [];
 
-  static get styles() {
-    return [
-      buttonStyle,
-      css`
-        :host {
-          display: block;
-          position: relative;
-          padding: 0 1rem;
-          min-height: 50px;
-          max-height: 200px;
-          overflow: auto;
-        }
-
-        p {
-          margin: 0 0 0.25rem;
-          font-family: var(--ave-monospace-font);
-          font-size: 0.875rem;
-          line-height: 1.5;
-        }
-
-        p:first-of-type {
-          margin-top: 1rem;
-        }
-
-        p:last-of-type {
-          margin-bottom: 1rem;
-        }
-      `
-    ];
+  protected createRenderRoot() {
+    return this;
   }
 
   protected render() {
@@ -79,7 +50,7 @@ export class ApiViewerDemoEvents extends LitElement {
         log.length
           ? renderEvents(log)
           : html`
-              <p>Interact with component to see the event log.</p>
+              <p class="event">Interact with component to see the event log.</p>
             `
       )}
     `;
