+++ src/api-viewer-header.ts
@@ -36,6 +36,13 @@ export class ApiViewerHeader extends LitElement {
           margin-bottom: 0.5rem;
         }
       }
+
+      ::slotted(api-viewer-toggle) {
+        display: block;
+        margin-right: 0.5rem;
+        color: var(--ave-header-color);
+        font-size: 0.875rem;
+      }
     `;
   }
