+++ src/api-viewer.ts
@@ -2,7 +2,6 @@ import {
   LitElement,
   html,
   customElement,
-  css,
   property,
   TemplateResult
 } from 'lit-element';
@@ -11,6 +10,7 @@ import { fetchJson } from './lib/fetch-json.js';
 import { ElementInfo } from './lib/types.js';
 import { queryTemplates } from './lib/utils.js';
 import './api-viewer-content.js';
+import styles from './api-viewer-styles.js';
 
 @customElement('api-viewer')
 export class ApiViewer extends LitElement {
@@ -50,36 +50,7 @@ export class ApiViewer extends LitElement {
   }
 
   static get styles() {
-    return css`
-      :host {
-        display: block;
-        box-sizing: border-box;
-        max-width: 800px;
-        min-width: 360px;
-        font-size: 1rem;
-        font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
-          'Oxygen-Sans', Ubuntu, Cantarell, sans-serif;
-        border: 1px solid var(--ave-border-color);
-        border-radius: var(--ave-border-radius);
-
-        --ave-primary-color: #01579b;
-        --ave-accent-color: #d63200;
-        --ave-border-color: rgba(0, 0, 0, 0.12);
-        --ave-border-radius: 4px;
-        --ave-header-color: #fff;
-        --ave-item-color: rgba(0, 0, 0, 0.87);
-        --ave-label-color: #424242;
-        --ave-link-color: #01579b;
-        --ave-link-hover-color: #d63200;
-        --ave-tab-color: rgba(0, 0, 0, 0.54);
-        --ave-monospace-font: Menlo, 'DejaVu Sans Mono', 'Liberation Mono',
-          Consolas, 'Courier New', monospace;
-      }
-
-      .warn {
-        padding: 1rem;
-      }
-    `;
+    return styles;
   }
 
   protected render() {
