--- src/lib/button-style.ts
@@ -1,20 +1,0 @@
-import { css } from 'lit-element';
-
-export default css`
-  button {
-    position: absolute;
-    top: 0.5rem;
-    right: 0.5rem;
-    text-transform: uppercase;
-    border: none;
-    border-radius: 0.25em;
-    cursor: pointer;
-    background: var(--ave-button-background, rgba(0, 0, 0, 0.3));
-    color: var(--ave-button-color, #fff);
-  }
-
-  button:focus,
-  button:hover {
-    background: var(--ave-button-active-background, rgba(0, 0, 0, 0.6));
-  }
-`;
